1. Оставлено одно активити, в котором располагается панель навигации и placeholder для фрагментов. 2. Главный экран и экран с деталями переведены на фрагменты.
