+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Base.Theme.FilmSearcher"
         tools:targetApi="31">
         <activity
-            android:name=".DetailsActivity"
+            android:name=".DetailsFragment"
             android:exported="false" />
         <activity
             android:name=".MainActivity"
