+++ app/src/main/res/values/strings.xml
@@ -11,4 +11,6 @@
     <string name="menu_selections_title">Подборки</string>
     <string name="menu_settings_title">Настройки</string>
     <string name="page_title">Film search</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
